+++ rx-java/src/main/java/io/vertx/rxjava/ext/sql/SQLClientHelper.java
@@ -11,6 +11,9 @@
 
 package io.vertx.rxjava.ext.sql;
 
+import io.vertx.rxjava.ext.sql.impl.InTransactionCompletable;
+import io.vertx.rxjava.ext.sql.impl.InTransactionObservable;
+import io.vertx.rxjava.ext.sql.impl.InTransactionSingle;
 import rx.Completable;
 import rx.Observable;
 import rx.Single;
@@ -25,6 +28,22 @@ import java.util.function.Function;
 public class SQLClientHelper {
 
   /**
+   * Creates a {@link Observable.Transformer} decorating an {@link Observable} with transaction management for a given {@link SQLConnection}.
+   * <p>
+   * If the upstream {@link Observable} completes (<em>onComplete</em>), the transaction is committed.
+   * If the upstream {@link Observable} emits an error (<em>onError</em>), the transaction is rollbacked.
+   * <p>
+   * Eventually, the given {@link SQLConnection} is put back in <em>autocommit</em> mode.
+   *
+   * @param sqlConnection the {@link SQLConnection} used for database operations and transaction management
+   * @param <T> the type of the items emitted by the upstream {@link Observable}
+   * @return a {@link Observable.Transformer} decorating an {@link Observable} with transaction management
+   */
+  public static <T> Observable.Transformer<T, T> txObservableTransformer(SQLConnection sqlConnection) {
+    return new InTransactionObservable<>(sqlConnection);
+  }
+
+  /**
    * Generates a {@link Observable} from {@link SQLConnection} operations executed inside a transaction.
    *
    * @param client the {@link SQLClient}
@@ -34,8 +53,24 @@ public class SQLClientHelper {
    */
   public static <T> Observable<T> inTransactionObservable(SQLClient client, Function<SQLConnection, Observable<T>> sourceSupplier) {
     return client.rxGetConnection().flatMapObservable(conn -> {
-      return sourceSupplier.apply(conn).compose(new InTransactionObservable<>(conn)).doAfterTerminate(conn::close);
+      return sourceSupplier.apply(conn).compose(txObservableTransformer(conn)).doAfterTerminate(conn::close);
     });
+  }
+
+  /**
+   * Creates a {@link Single.Transformer} decorating a {@link Single} with transaction management for a given {@link SQLConnection}.
+   * <p>
+   * If the upstream {@link Single} emits a value (<em>onSuccess</em>), the transaction is committed.
+   * If the upstream {@link Single} emits an error (<em>onError</em>), the transaction is rollbacked.
+   * <p>
+   * Eventually, the given {@link SQLConnection} is put back in <em>autocommit</em> mode.
+   *
+   * @param sqlConnection the {@link SQLConnection} used for database operations and transaction management
+   * @param <T> the type of the item emitted by the upstream {@link Single}
+   * @return a {@link Single.Transformer} decorating a {@link Single} with transaction management
+   */
+  public static <T> Single.Transformer<T, T> txSingleTransformer(SQLConnection sqlConnection) {
+    return new InTransactionSingle<>(sqlConnection);
   }
 
   /**
@@ -48,8 +83,23 @@ public class SQLClientHelper {
    */
   public static <T> Single<T> inTransactionSingle(SQLClient client, Function<SQLConnection, Single<T>> sourceSupplier) {
     return client.rxGetConnection().flatMap(conn -> {
-      return sourceSupplier.apply(conn).compose(new InTransactionSingle<>(conn)).doAfterTerminate(conn::close);
+      return sourceSupplier.apply(conn).compose(txSingleTransformer(conn)).doAfterTerminate(conn::close);
     });
+  }
+
+  /**
+   * Creates a {@link Completable.Transformer} decorating a {@link Completable} with transaction management for a given {@link SQLConnection}.
+   * <p>
+   * If the upstream {@link Completable} completes (<em>onComplete</em>), the transaction is committed.
+   * If the upstream {@link Completable} emits an error (<em>onError</em>), the transaction is rollbacked.
+   * <p>
+   * Eventually, the given {@link SQLConnection} is put back in <em>autocommit</em> mode.
+   *
+   * @param sqlConnection the {@link SQLConnection} used for database operations and transaction management
+   * @return a {@link Completable.Transformer} decorating a {@link Completable} with transaction management
+   */
+  public static Completable.Transformer txCompletableTransformer(SQLConnection sqlConnection) {
+    return new InTransactionCompletable(sqlConnection);
   }
 
   /**
@@ -61,7 +111,7 @@ public class SQLClientHelper {
    */
   public static Completable inTransactionCompletable(SQLClient client, Function<SQLConnection, Completable> sourceSupplier) {
     return client.rxGetConnection().flatMapCompletable(conn -> {
-      return sourceSupplier.apply(conn).compose(new InTransactionCompletable(conn)).doAfterTerminate(conn::close);
+      return sourceSupplier.apply(conn).compose(txCompletableTransformer(conn)).doAfterTerminate(conn::close);
     });
   }
