+++ rx-java/src/test/java/io/vertx/rxjava/ext/sql/InTransactionSingleTest.java
@@ -45,7 +45,7 @@ public class InTransactionSingleTest extends SQLTestBase {
         .andThen(uniqueNames(conn))
         .<List<String>>collect(ArrayList::new, List::add).toSingle()
         .compose(upstream -> e == null ? upstream : upstream.flatMap(names -> Single.error(e)))
-        .compose(new InTransactionSingle<>(conn))
+        .compose(SQLClientHelper.txSingleTransformer(conn))
         .flatMap(names -> rxAssertAutoCommit(conn).andThen(Single.just(names)))
         .doAfterTerminate(conn::close);
     });
