+++ rx-java2/src/main/java/io/vertx/reactivex/ext/sql/impl/InTransactionObservable.java
@@ -9,20 +9,14 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.reactivex.ext.sql;
+package io.vertx.reactivex.ext.sql.impl;
 
 import io.reactivex.Observable;
 import io.reactivex.ObservableSource;
 import io.reactivex.ObservableTransformer;
+import io.vertx.reactivex.ext.sql.SQLConnection;
 
 /**
- * Decorates a {@link Observable} with transaction management for a given {@link SQLConnection}.
- * <p>
- * If the {@link Observable} completes (<em>onComplete</em>), the transaction is committed.
- * If the {@link Observable} emits an error (<em>onError</em>), the transaction is rollbacked.
- * <p>
- * Eventually, the given {@link SQLConnection} is put back in <em>autocommit</em> mode.
- *
  * @author Thomas Segismont
  */
 public class InTransactionObservable<T> implements ObservableTransformer<T, T> {
